move code related to default connection to another file to fix code climate issue
